backend/requirements.txt
drf-yasg
django
djangorestframework
djangorestframework-simplejwt
django-allauth
django-otp
django-rest-auth
channels
channels-redis
celery
redis
psycopg2-binary
twilio
reportlab
qrcode
pillow
requests
pandas
asgiref
PyJWT
python-dateutil
sqlparse
packaging
cryptography
colorama
click
attrs
yarl
multidict
amqp
kombu
billiard
prompt_toolkit
msgpack
six
tzdata
wcwidth
